feat: :passport_control: implemented permission based access control for goals on the backend

+++ apps/app1-bff/src/app/goal/goal.controller.ts
@@ -1,12 +1,15 @@
+import { subject } from '@casl/ability';
+import { CaslAbility, UseCasl, UseDtoUserId } from '@goalie/nest-auth';
 import { CreateGoalDto, Goal, GoalActivity } from '@goalie/shared/goals';
-import { Body, Controller, Get, HttpStatus, Param, Post } from '@nestjs/common';
+import { Body, Controller, ForbiddenException, Get, HttpStatus, Param, Post } from '@nestjs/common';
 import { ApiResponse, ApiTags } from '@nestjs/swagger';
-import { Observable } from 'rxjs';
-import { UseDtoUserId } from '../auth/dto-user-id.interceptor';
+import { map, mergeMap, Observable } from 'rxjs';
+import { AppAbility } from '../auth/casl/casl.factory';
 import { ReqUser } from '../auth/request-user.decorator';
 import { User } from '../auth/user.model';
 import { GoalsService } from './goals.service';
 
+@UseCasl()
 @ApiTags('goals')
 @Controller('goals')
 export class GoalController {
@@ -24,7 +27,11 @@ export class GoalController {
     description: 'The goal has been successfully created.',
     type: String,
   })
-  public create(@Body() createGoalDto: CreateGoalDto): Observable<string> {
+  public create(@CaslAbility() ability: AppAbility, @Body() createGoalDto: CreateGoalDto): Observable<string> {
+    if (ability.cannot('create', 'Goal')) {
+      throw new ForbiddenException(`You are not allowed to create a goal.`);
+    }
+
     return this.service.create(createGoalDto);
   }
 
@@ -39,8 +46,16 @@ export class GoalController {
     description: 'The list of goals has been successfully retrieved.',
     type: [Goal],
   })
-  public getAll(@ReqUser() user: User): Observable<Goal[]> {
-    return this.service.getAll(user.userId);
+  public getAll(@CaslAbility() ability: AppAbility, @ReqUser() user: User): Observable<Goal[]> {
+    return this.service.getAll(user.userId).pipe(
+      map((goals) => {
+        if (goals.some((goal) => ability.cannot('read', subject('Goal', goal)))) {
+          throw new ForbiddenException(`You are not allowed to access some of the goals.`);
+        }
+
+        return goals;
+      })
+    );
   }
 
   /**
@@ -55,8 +70,16 @@ export class GoalController {
     description: 'The goal has been successfully retrieved.',
     type: Goal,
   })
-  public get(@ReqUser() _user: User, @Param('id') id: string): Observable<Goal> {
-    return this.service.get(id);
+  public get(@CaslAbility() ability: AppAbility, @Param('id') id: string): Observable<Goal> {
+    return this.service.get(id).pipe(
+      map((goal) => {
+        if (ability.cannot('read', subject('Goal', goal))) {
+          throw new ForbiddenException(`You are not allowed to access this goal.`);
+        }
+
+        return goal;
+      })
+    );
   }
 
   /**
@@ -71,8 +94,16 @@ export class GoalController {
     description: 'The goal activity has been successfully retrieved.',
     type: [GoalActivity],
   })
-  public getActivity(@ReqUser() _user: User, @Param('id') id: string): Observable<GoalActivity[]> {
-    return this.service.getActivity(id);
+  public getActivity(@CaslAbility() ability: AppAbility, @Param('id') id: string): Observable<GoalActivity[]> {
+    return this.service.get(id).pipe(
+      mergeMap((goal) => {
+        if (ability.cannot('read', subject('Goal', goal))) {
+          throw new ForbiddenException(`You are not allowed to access this goal.`);
+        }
+
+        return this.service.getActivity(id);
+      })
+    );
   }
 
   /**
@@ -86,8 +117,16 @@ export class GoalController {
     status: HttpStatus.OK,
     description: 'The goal has been successfully completed.',
   })
-  public complete(@ReqUser() _user: User, @Param('id') id: string): Observable<unknown> {
-    return this.service.complete(id);
+  public complete(@CaslAbility() ability: AppAbility, @Param('id') id: string): Observable<unknown> {
+    return this.service.get(id).pipe(
+      mergeMap((goal) => {
+        if (ability.cannot('update', subject('Goal', goal))) {
+          throw new ForbiddenException(`You are not allowed to complete this goal.`);
+        }
+
+        return this.service.complete(id);
+      })
+    );
   }
 
   /**
@@ -101,7 +140,15 @@ export class GoalController {
     status: HttpStatus.OK,
     description: 'The goal has been successfully deleted.',
   })
-  public delete(@ReqUser() _user: User, @Param('id') id: string): Observable<unknown> {
-    return this.service.delete(id);
+  public delete(@CaslAbility() ability: AppAbility, @Param('id') id: string): Observable<unknown> {
+    return this.service.get(id).pipe(
+      mergeMap((goal) => {
+        if (ability.cannot('delete', subject('Goal', goal))) {
+          throw new ForbiddenException(`You are not allowed to delete this goal.`);
+        }
+
+        return this.service.delete(id);
+      })
+    );
   }
 }
